requirements-dev.txt
-r requirements.txt

# enforcing dev practices on commit
pre-commit==2.19.0

# tests
pytest-django
